Move AudioOutputWorker to a separate file
